edhoc: refactor c wrapper and make sure no_std example still works
