+++ lib/src/lib.rs
@@ -39,6 +39,9 @@ mod edhoc;
 ))]
 use edhoc::*;
 
+mod c_wrapper;
+use c_wrapper::*;
+
 #[cfg(any(
     feature = "hacspec-hacspec",
     feature = "hacspec-cc2538",
@@ -335,24 +338,6 @@ mod rust {
         cred_r: &'a str,    // R's full credential
     }
 
-
-    #[repr(C)]
-    pub struct RustEdhocInitiatorC {
-        state: State,       // opaque state
-        i: *const u8,
-        i_len: usize,
-        g_r: *const u8,
-        g_r_len: usize,
-        id_cred_i: *const u8,
-        id_cred_i_len: usize,
-        cred_i: *const u8,
-        cred_i_len: usize,
-        id_cred_r: *const u8,
-        id_cred_r_len: usize,
-        cred_r: *const u8,
-        cred_r_len: usize,
-    }
-
     #[derive(Default, Copy, Clone, Debug)]
     pub struct RustEdhocResponder<'a> {
         state: State,       // opaque state
@@ -362,23 +347,6 @@ mod rust {
         cred_i: &'a str,    // I's full credential
         id_cred_r: &'a str, // identifier of R's credential
         cred_r: &'a str,    // R's full credential
-    }
-
-    #[repr(C)]
-    pub struct RustEdhocResponderC {
-        state: State,       // opaque state
-        r: *const u8,
-        r_len: usize,
-        g_i: *const u8,
-        g_i_len: usize,
-        id_cred_i: *const u8,
-        id_cred_i_len: usize,
-        cred_i: *const u8,
-        cred_i_len: usize,
-        id_cred_r: *const u8,
-        id_cred_r_len: usize,
-        cred_r: *const u8,
-        cred_r_len: usize,
     }
 
     impl<'a> RustEdhocResponder<'a> {
@@ -397,18 +365,6 @@ mod rust {
                 id_cred_r_len: self.id_cred_r.len(),
                 cred_r: self.cred_r.as_ptr(),
                 cred_r_len: self.cred_r.len(),
-            }
-        }
-
-        pub fn from_c(responder: &RustEdhocResponderC) -> RustEdhocResponder<'a> {
-            RustEdhocResponder {
-                state: responder.state,
-                r: unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(responder.r, responder.r_len)) },
-                g_i: unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(responder.g_i, responder.g_i_len)) },
-                id_cred_i: unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(responder.id_cred_i, responder.id_cred_i_len)) },
-                cred_i: unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(responder.cred_i, responder.cred_i_len)) },
-                id_cred_r: unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(responder.id_cred_r, responder.id_cred_r_len)) },
-                cred_r: unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(responder.cred_r, responder.cred_r_len)) },
             }
         }
 
@@ -531,18 +487,6 @@ mod rust {
                 id_cred_r_len: self.id_cred_r.len(),
                 cred_r: self.cred_r.as_ptr(),
                 cred_r_len: self.cred_r.len(),
-            }
-        }
-
-        pub fn from_c(initiator: &RustEdhocInitiatorC) -> RustEdhocInitiator<'a> {
-            RustEdhocInitiator {
-                state: initiator.state,
-                i: unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(initiator.i, initiator.i_len)) },
-                g_r: unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(initiator.g_r, initiator.g_r_len)) },
-                id_cred_i: unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(initiator.id_cred_i, initiator.id_cred_i_len)) },
-                cred_i: unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(initiator.cred_i, initiator.cred_i_len)) },
-                id_cred_r: unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(initiator.id_cred_r, initiator.id_cred_r_len)) },
-                cred_r: unsafe { core::str::from_utf8_unchecked(core::slice::from_raw_parts(initiator.cred_r, initiator.cred_r_len)) },
             }
         }
 
@@ -842,222 +786,3 @@ mod test {
     }
 }
 
-
-use core::{slice, str};
-use crate::rust::*;
-
-// Rust requires a panic handler in order to compile for cortex-m in no_std mode
-#[panic_handler]
-fn my_panic(_info: &core::panic::PanicInfo) -> ! {
-    loop {}
-}
-
-#[no_mangle]
-pub extern "C" fn edhoc_add(a: i32, b: i32) -> i32 {
-    a + b
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn responder_new(
-    r: *const u8,
-    r_len: usize,
-    g_i: *const u8,
-    g_i_len: usize,
-    id_cred_i: *const u8,
-    id_cred_i_len: usize,
-    cred_i: *const u8,
-    cred_i_len: usize,
-    id_cred_r: *const u8,
-    id_cred_r_len: usize,
-    cred_r: *const u8,
-    cred_r_len: usize,
-) -> RustEdhocResponderC {
-    RustEdhocResponder::new(
-        EdhocState::default(),
-        str::from_utf8_unchecked(slice::from_raw_parts(r, r_len)),
-        str::from_utf8_unchecked(slice::from_raw_parts(g_i, g_i_len)),
-        str::from_utf8_unchecked(slice::from_raw_parts(id_cred_i, id_cred_i_len)),
-        str::from_utf8_unchecked(slice::from_raw_parts(cred_i, cred_i_len)),
-        str::from_utf8_unchecked(slice::from_raw_parts(id_cred_r, id_cred_r_len)),
-        str::from_utf8_unchecked(slice::from_raw_parts(cred_r, cred_r_len)),
-    ).to_c()
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn initiator_new(
-    i: *const u8,
-    i_len: usize,
-    g_r: *const u8,
-    g_r_len: usize,
-    id_cred_i: *const u8,
-    id_cred_i_len: usize,
-    cred_i: *const u8,
-    cred_i_len: usize,
-    id_cred_r: *const u8,
-    id_cred_r_len: usize,
-    cred_r: *const u8,
-    cred_r_len: usize,
-) -> RustEdhocInitiatorC {
-    RustEdhocInitiator::new(
-        EdhocState::default(),
-        str::from_utf8_unchecked(slice::from_raw_parts(i, i_len)),
-        str::from_utf8_unchecked(slice::from_raw_parts(g_r, g_r_len)),
-        str::from_utf8_unchecked(slice::from_raw_parts(id_cred_i, id_cred_i_len)),
-        str::from_utf8_unchecked(slice::from_raw_parts(cred_i, cred_i_len)),
-        str::from_utf8_unchecked(slice::from_raw_parts(id_cred_r, id_cred_r_len)),
-        str::from_utf8_unchecked(slice::from_raw_parts(cred_r, cred_r_len)),
-    ).to_c()
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn initiator_prepare_message_1(
-    initiator_c: *mut RustEdhocInitiatorC,
-    message_1: *mut EdhocMessageBuffer,
-) -> i8 {
-    let mut initiator = RustEdhocInitiator::from_c(&*initiator_c);
-
-    let result = match initiator.prepare_message_1() {
-        Ok(msg_1) => {
-            *message_1 = msg_1;
-            0
-        },
-        Err(_) => -1,
-    };
-
-    *initiator_c = initiator.to_c();
-
-    result
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn responder_process_message_1(
-    responder_c: *mut RustEdhocResponderC,
-    message_1: *const EdhocMessageBuffer,
-) -> i8 {
-    let mut responder = RustEdhocResponder::from_c(&*responder_c);
-
-    let result = match responder.process_message_1(&*message_1) {
-        Ok(_) => 0,
-        Err(_) => -1,
-    };
-
-    *responder_c = responder.to_c();
-
-    result
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn responder_prepare_message_2(
-    responder_c: *mut RustEdhocResponderC,
-    message_2: *mut EdhocMessageBuffer,
-    c_r: *mut u8,
-) -> i8 {
-    let mut responder = RustEdhocResponder::from_c(&*responder_c);
-
-    let result = match responder.prepare_message_2() {
-        Ok((msg_2, c_r_res)) => {
-            *message_2 = msg_2;
-            *c_r = c_r_res;
-            0
-        },
-        Err(_) => -1,
-    };
-
-    *responder_c = responder.to_c();
-
-    result
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn initiator_process_message_2(
-    initiator_c: *mut RustEdhocInitiatorC,
-    message_2: *const EdhocMessageBuffer,
-    c_r: *mut u8,
-) -> i8 {
-    let mut initiator = RustEdhocInitiator::from_c(&*initiator_c);
-
-    let result = match initiator.process_message_2(&*message_2) {
-        Ok(c_r_res) => {
-            *c_r = c_r_res;
-            0
-        },
-        Err(_) => -1,
-    };
-
-    *initiator_c = initiator.to_c();
-
-    result
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn initiator_prepare_message_3(
-    initiator_c: *mut RustEdhocInitiatorC,
-    message_3: *mut EdhocMessageBuffer,
-    prk_out: *mut [u8; SHA256_DIGEST_LEN],
-) -> i8 {
-    let mut initiator = RustEdhocInitiator::from_c(&*initiator_c);
-
-    let result = match initiator.prepare_message_3() {
-        Ok((msg_3, prk_out_res)) => {
-            *message_3 = msg_3;
-            *prk_out = prk_out_res;
-            0
-        },
-        Err(_) => -1,
-    };
-
-    *initiator_c = initiator.to_c();
-
-    result
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn responder_process_message_3(
-    responder_c: *mut RustEdhocResponderC,
-    message_3: *const EdhocMessageBuffer,
-    prk_out: *mut [u8; SHA256_DIGEST_LEN],
-) -> i8 {
-    let mut responder = RustEdhocResponder::from_c(&*responder_c);
-
-    let result = match responder.process_message_3(&*message_3) {
-        Ok(prk_out_res) => {
-            *prk_out = prk_out_res;
-            0
-        },
-        Err(_) => -1,
-    };
-
-    *responder_c = responder.to_c();
-
-    result
-}
-
-#[cfg(test)]
-mod test_c {
-    use super::*;
-
-    #[test]
-    fn test_new_responder() {
-        const ID_CRED_I: &[u8] = "a104412b".as_bytes();
-        const ID_CRED_R: &[u8] = "a104410a".as_bytes();
-        const CRED_I: &[u8] = "A2027734322D35302D33312D46462D45462D33372D33322D333908A101A5010202412B2001215820AC75E9ECE3E50BFC8ED60399889522405C47BF16DF96660A41298CB4307F7EB62258206E5DE611388A4B8A8211334AC7D37ECB52A387D257E6DB3C2A93DF21FF3AFFC8".as_bytes();
-        const G_I: &[u8] = "ac75e9ece3e50bfc8ed60399889522405c47bf16df96660a41298cb4307f7eb6".as_bytes();
-        const CRED_R: &[u8] = "A2026008A101A5010202410A2001215820BBC34960526EA4D32E940CAD2A234148DDC21791A12AFBCBAC93622046DD44F02258204519E257236B2A0CE2023F0931F1F386CA7AFDA64FCDE0108C224C51EABF6072".as_bytes();
-        const R: &[u8] = "72cc4761dbd4c78f758931aa589d348d1ef874a7e303ede2f140dcf3e6aa4aac".as_bytes();
-
-        let resp = unsafe { responder_new(
-            R.as_ptr(),
-            R.len(),
-            G_I.as_ptr(),
-            G_I.len(),
-            ID_CRED_I.as_ptr(),
-            ID_CRED_I.len(),
-            CRED_I.as_ptr(),
-            CRED_I.len(),
-            ID_CRED_R.as_ptr(),
-            ID_CRED_R.len(),
-            CRED_R.as_ptr(),
-            CRED_R.len(),
-        ) };
-    }
-}
